+++ packages/search/docs/stories/GlobalSearch.story.js
@@ -3,6 +3,7 @@ import styled from 'styled-components';
 import { RainbowFirebaseApp } from '@rainbow-modules/app';
 import GlobalSearch from '../../src/components/GlobalSearch';
 import SearchEntity from '../../src/components/SearchEntity';
+import useLocalRecentSearches from '../../src/hooks/useLocalRecentSearches';
 import searchBooksAlgolia from './algolia/searchBooksAlgolia';
 import searchAuthorsAlgolia from './algolia/searchAuthorsAlgolia';
 import searchMoviesTitleMongo from './mongo/searchMoviesTitleMongo';
@@ -38,6 +39,7 @@ export const AlgoliaGlobalSearch = () => {
 };
 
 export const MongoAtlasGlobalSearch = () => {
+    const [recents, addRecents] = useLocalRecentSearches('mongo-atlas-recent-searches');
     return (
         <RainbowFirebaseApp>
             <StyledGlobalSearch
@@ -45,6 +47,8 @@ export const MongoAtlasGlobalSearch = () => {
                 placeholder="Search"
                 // eslint-disable-next-line no-alert
                 onSelect={(item) => alert(JSON.stringify(item))}
+                recents={recents}
+                onSearch={({ query }) => addRecents(query)}
             >
                 <SearchEntity name="Movies" onAutocomplete={searchMoviesTitleMongo} />
                 <SearchEntity name="Plot" onAutocomplete={searchMoviesPlotMongo} />